// cypress/enums/global.ts
export enum GlobalDataTestIds {
  CUSTOM_TABLE_WRAPPER = 'custom-table-wrapper',
  MARKETPLACE_DROPDOWN = 'marketplace-dropdown',
  MULTI_SELECT_DROPDOWN = 'multi-select-dropdown',
  TOAST_WRAPPER = 'toastWrapper',
  EMPTY_PLACEHOLDER = 'empty-placeholder',
  SEARCH_WRAPPER = 'search-wrapper',
  DOWNLOAD_BUTTON = 'download-button',
  ADDED_FILTERS_TAB = 'added-filters-tab',
  SINGLE_FILTER_CONTAINER = 'single-filter-container',
  CLEAR_BUTTON = 'clear-button',
  ADDED_FILTERS_CONTAINER = 'added-filters-container',
  DOWNLOAD_OPTIONS_CONTAINER = 'download_options-container',
  UPLOAD_FILE_DIALOG = 'upload-file-dialog',
  UPLOAD_BUTTON = 'upload-button',
  PRODUCT_DATA = 'product-data',
  PRODUCT_NAME = 'product-name',
  PRODUCT_META_DATA = 'product-meta-data',
  PRODUCT_CONTAINER = 'product-container',
  CONFIRMATION_POPUP = 'confirmation-popup',
  CONFIRMATION_POPUP_CANCEL_ICON_BUTTON = 'confirmation-cancel-icon-button',
  CONFIRMATION_POPUP_CANCEL_ACTION_BUTTON = 'confirmation-cancel-action-button',
  CONFIRMATION_POPUP_CONFIRM_ACTION_BUTTON = 'confirmation-confirm-action-button',
}