// cypress/enums/keyword-tracker.ts
export enum KeywordTrackerDataTestIds {
  KEYWORD_TRACKER_TABS = 'kt-tabs-select',
  TABLE_BUTTON_CONTAINER = 'kt-tableButtonContainer',
  ADD_KEYWORD_DIALOG = 'add-keyword-dialog',
  ENTER_KEYWORDS_INPUT = 'enter-keywords',
  BULK_UPLOAD_CONTAINER = 'bulk-upload-container',
  DOWNLOAD_FILE_CONTAINER = 'download-file-container',
  STATUS_CONFIRMATION_POPUP = 'status-confirmation',
}